requirements/requirements.txt
easydict>=1.9
numpy>=1.21.4
pillow~=8.4.0
pyyaml~=6.0
sklearn~=0.0
scikit-learn~=0.24.2
torchvision~=0.11.2
numba~=0.53.1
opencv-python>=4.5.4.58
pandas>=1.3.4
tqdm~=4.64.1
h5py>=3.6.0
matplotlib>=3.5.0
pip>=21.2.4
wheel>=0.37.0
brotli~=1.0.9
tornado~=6.1
setuptools>=59.2.0
future~=0.18.2
click>=8.1.3
xlsxwriter>=3.0.3
typing_extensions>=3.10.0.2
llvmlite < 0.37.0
cffi>=1.15.1
colorama==0.4.6
pygments==2.14.0
psutil==5.9.4
scipy==1.10.0
jinja2>=3.1.2
ipykernel>=6.21.0
nbformat>=5.7.3
pytz>=2021.3
openpyxl>=3.0.10
markupsafe>=2.1.2
python-dateutil>=2.8.2
numexpr>=2.8.1
xlrd>=1.1.0
fonttools>=4.28.1
path>=16.2.0
cycler>=0.11.0
packaging>=21.3
ipython>=7.29.0
pyparsing>=3.0.6
kiwisolver>=1.3.2
requests>=2.28.2
keras>=2.7.0
ipywidgets>=8.0.4
notebook>=6.5.2
joblib>=1.1.0
cloudpickle>=2.0.0
six>=1.16.0
zipp>=3.6.0
defusedxml>=0.7.1
astunparse>=1.6.3
dill>=0.3.4
tensorboard>=2.7.0
tensorflow>=2.7.0
tensorflow-gpu>=2.7.0
urllib3>=1.26.7
threadpoolctl>=2.2.0
image>=1.5.33